Refactor and update size limit
