+++ src/components/landing-page/Features.tsx
@@ -14,7 +14,7 @@ type Feature = {
 const features: Array<Feature> = [
   {
     title: 'Store everything',
-    description: 'You cannot only store your trip reports, but you can also upload images, add tags to your routes, pin POIs on an interactive map and much more.',
+    description: 'You cannot only store your trip reports, but you can also upload geo-referenced images, add GPX tracks and categories to your tours and much more.',
     icon: <FeatureStorage className='feature-icon'/>
   },
   {
